feat: Add Gradle Wrapper for Android
This commit introduces the Gradle Wrapper files for the Android part of the project. This includes:

- `gradle-wrapper.properties`: Configures the Gradle distribution URL and other wrapper settings.
- `gradlew`: The Gradle Wrapper script for Unix-based systems.
- `gradlew.bat`: The Gradle Wrapper script for Windows.
- `gradle-wrapper.jar`: The Gradle Wrapper JAR file.

Adding the Gradle Wrapper ensures a consistent Gradle version is used for building the Android project across different development environments.

Additionally, `devtools_options.yaml` files have been added to the root directory and the `example` directory. These files are used to store settings for Dart & Flutter DevTools.